+++ mail/mail-uniffi/src/mail/datatypes.rs
@@ -701,35 +701,31 @@ impl From<RealSpamAction> for SpamAction {
     }
 }
 
-/// TODO: Document this enum.
+/// See [`RealSwipeAction`]
 #[derive(Clone, Copy, Debug, Default, Eq, Hash, PartialEq, UniffiEnum)]
-#[repr(u8)]
 pub enum SwipeAction {
-    /// TODO: Document this variant.
-    Trash = 0,
-
-    /// TODO: Document this variant.
-    Spam = 1,
-
-    /// TODO: Document this variant.
-    Star = 2,
-
-    /// TODO: Document this variant.
+    NoAction,
+    Trash,
+    Spam,
+    Star,
     #[default]
-    Archive = 3,
-
-    /// TODO: Document this variant.
-    MarkAsRead = 4,
+    Archive,
+    MarkAsRead,
+    LabelAs,
+    MoveTo,
 }
 
 impl From<SwipeAction> for RealSwipeAction {
     fn from(value: SwipeAction) -> Self {
         match value {
+            SwipeAction::NoAction => RealSwipeAction::NoAction,
             SwipeAction::Trash => RealSwipeAction::Trash,
             SwipeAction::Spam => RealSwipeAction::Spam,
             SwipeAction::Star => RealSwipeAction::Star,
             SwipeAction::Archive => RealSwipeAction::Archive,
             SwipeAction::MarkAsRead => RealSwipeAction::MarkAsRead,
+            SwipeAction::LabelAs => RealSwipeAction::LabelAs,
+            SwipeAction::MoveTo => RealSwipeAction::MoveTo,
         }
     }
 }
@@ -737,11 +733,14 @@ impl From<SwipeAction> for RealSwipeAction {
 impl From<RealSwipeAction> for SwipeAction {
     fn from(value: RealSwipeAction) -> Self {
         match value {
+            RealSwipeAction::NoAction => SwipeAction::NoAction,
             RealSwipeAction::Trash => SwipeAction::Trash,
             RealSwipeAction::Spam => SwipeAction::Spam,
             RealSwipeAction::Star => SwipeAction::Star,
             RealSwipeAction::Archive => SwipeAction::Archive,
             RealSwipeAction::MarkAsRead => SwipeAction::MarkAsRead,
+            RealSwipeAction::LabelAs => SwipeAction::LabelAs,
+            RealSwipeAction::MoveTo => SwipeAction::MoveTo,
         }
     }
 }